3% Progress??
The MapGenerator script no longer spawns players (no long hardcoded) and it no longer sets the material of the spawned map as the spawned map is now a blueprint. Overall, slight changes to make the map be ready for multiplayer.

Fixed a bit of the config folder due to updated gitignore which removed the folder. Major fixes on it still needed however.