New classes and foundational code changes. Read description.
Added a class called DoStatic, literally just a class full of static methods, makes things easier if a method can be static so that any class in the future can just include the file and use any static methods created for the project.

PlayerCharacter now needs a MapGenerator making the MapGenerator mandatory for all levels and there should only be one.