Mostly successful threading of Generating a new chunk. Read Description...
This method that was implemented works if the chunkRadius is not massive (which it should not be).  The method first spawns the new chunk and then the chunk will create a thread to generate itself. This might cause a problem if the chunkradius is massive as having too many new chunks means too many threads which ultimately leads to worse performance.

A new solution might involve only one thread but for this commit, this is a working version.